pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.7",
    "fastapi-users>=14.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]
